Extract Connection to subclass for #63
