+++ Core/src/main/java/orpheus/core/net/SocketConnectionListener.java
@@ -49,8 +49,7 @@ public class SocketConnectionListener {
                     client.getPort()
                 ));
 
-                var connection = Connection.forSocket(client);
-                server.connectTo(connection);
+                server.connectTo(SocketConnection.forSocket(client));
             } catch (SocketTimeoutException ste) {
                 // this is not an error, it just means no client has attempted to connect
             } catch (IOException ex) {
